[Invoice][CuongHLT]update create/edit invoice

+++ src/main/java/com/example/retro_care/customer/controller/CustomerController.java
@@ -1,6 +1,8 @@
 package com.example.retro_care.customer.controller;
 
+import com.example.retro_care.customer.dto.CreatCode;
 import com.example.retro_care.customer.dto.CustomerDto;
+import com.example.retro_care.customer.dto.ICustomerDto;
 import com.example.retro_care.customer.model.Customer;
 import com.example.retro_care.customer.service.ICustomerService;
 import org.springframework.beans.BeanUtils;
@@ -12,7 +14,12 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
 import org.springframework.validation.BindingResult;
+import org.springframework.validation.FieldError;
 import org.springframework.web.bind.annotation.*;
+
+import javax.validation.Valid;
+import java.util.HashMap;
+import java.util.Map;
 
 @Controller
 @CrossOrigin("*")
@@ -24,13 +31,23 @@ public class CustomerController {
     /**
      * Author: TinDT
      * Goal: get customer for page to create customer information
-     * return creation page  of customers
+     * return HttpStatus
      */
     @GetMapping("/dto/create")
     public ResponseEntity<CustomerDto> getCustomerForCreate() {
         CustomerDto customerDto = new CustomerDto();
+        String customerCode = CreatCode.generateCustomerCode();
+        while (true) {
+            if (customerService.findCustomerByCode(customerCode) == null) {
+                break;
+            }
+        }
+        customerDto.setCode(customerCode);
+        System.out.println(customerDto.getCode());
         return new ResponseEntity<>(customerDto, HttpStatus.OK);
     }
+
+
 
     /**
      * Author: TinDT
@@ -38,19 +55,20 @@ public class CustomerController {
      * * return HttpStatus
      */
     @PostMapping("/create")
-    public ResponseEntity<Customer> saveCustomer(@RequestBody CustomerDto customerDto, BindingResult bindingResult) {
+    public ResponseEntity<?> saveCustomer(@Valid  @RequestBody CustomerDto customerDto, BindingResult bindingResult) {
         Customer customer = new Customer();
         Customer saveCustomer;
         new CustomerDto().validate(customerDto, bindingResult);
         if (bindingResult.hasErrors()) {
-            return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+            Map<String,String> errors = new HashMap<>();
+            for (FieldError err: bindingResult.getFieldErrors()) {
+                errors.put(err.getField(), err.getDefaultMessage());
+            }
+            return new ResponseEntity<>(errors,HttpStatus.NOT_ACCEPTABLE);
         }
         BeanUtils.copyProperties(customerDto, customer);
         saveCustomer = customerService.saveCustomer(customer);
-        if (saveCustomer != null) {
-            return new ResponseEntity<>(saveCustomer, HttpStatus.OK);
-        }
-        return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+            return new ResponseEntity<>("Thêm mới khách hàng thành công", HttpStatus.OK);
 
     }
 
@@ -59,11 +77,22 @@ public class CustomerController {
      * Goal: update information of customer
      * * return HttpStatus
      */
-    @PatchMapping("/update")
-    public ResponseEntity<Customer> updateCustomer(@RequestBody Customer customer) {
+    @PutMapping("/update")
+    public ResponseEntity<?> updateCustomer(@Valid @RequestBody CustomerDto customerDto,BindingResult bindingResult) {
+        new CustomerDto().validate(customerDto,bindingResult);
+        if (bindingResult.hasErrors()) {
+            Map<String, String> errors = new HashMap<>();
+            for (FieldError err: bindingResult.getFieldErrors()) {
+                errors.put(err.getField(), err.getDefaultMessage());
+            }
+            return new ResponseEntity<>(errors, HttpStatus.NOT_ACCEPTABLE);
+        }
+        Customer customer = new Customer();
+        BeanUtils.copyProperties(customerDto,customer);
         customerService.updateCustomer(customer);
-        return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+        return new ResponseEntity<>("Cập nhật thông tin khách hành thành công",HttpStatus.OK);
     }
+
 
     /**
      * Author: TinDT
@@ -71,12 +100,11 @@ public class CustomerController {
      * * return HttpStatus
      */
     @GetMapping("/{id}")
-    public ResponseEntity<Customer> detailCustomer(@PathVariable Long id) {
+    public ResponseEntity<?> detailCustomer(@PathVariable Long id) {
         Customer customer = customerService.findCustomerById(id);
-        if (customer != null) {
-            return new ResponseEntity<>(customer, HttpStatus.OK);
-        }
-        return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        CustomerDto customerDto = new CustomerDto();
+        BeanUtils.copyProperties(customer, customerDto);
+        return new ResponseEntity<>(customerDto, HttpStatus.OK);
     }
 
     /**
@@ -89,11 +117,22 @@ public class CustomerController {
                                              @RequestParam(defaultValue = "", required = false) String search,
                                              @RequestParam(defaultValue = "", required = false) String code,
                                              @RequestParam(defaultValue = "", required = false) String address,
-                                             @RequestParam(defaultValue = "", required = false) String groupValue,
-                                             @RequestParam(defaultValue = "code", required = false) String sortItem) {
+                                             @RequestParam(defaultValue = "" ) String groupValue,
+                                             @RequestParam(defaultValue = "" ) String sortItem) {
         Pageable pageable = PageRequest.of(page, 5);
-        Page<Customer> customers = customerService.findAllCustomer(pageable, "%" + search + "%", "%" + code + "%", "%" + address + "%", groupValue, sortItem);
-        if (!customers.isEmpty()) {
+        Page<ICustomerDto> customers = customerService.findAllCustomer("%" + search + "%", "%" + code + "%", "%" + address + "%", groupValue, sortItem, pageable);
+        if (customers.getTotalElements()!=0) {
+            return new ResponseEntity<>(customers, HttpStatus.OK);
+        }
+        return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+    }
+    @GetMapping("/list-customer")
+    public ResponseEntity<?> getAll(@RequestParam(defaultValue = "0", required = false)int page,
+                                    @RequestParam(defaultValue = "", required = false) String searchName) {
+        Pageable pageable = PageRequest.of(page,5);
+        Page<Customer> customers = customerService.findAllByName(pageable, searchName);
+        System.out.println(customers.getContent());
+        if (customers.getTotalElements()!=0) {
             return new ResponseEntity<>(customers, HttpStatus.OK);
         }
         return new ResponseEntity<>(HttpStatus.NO_CONTENT);
@@ -106,6 +145,10 @@ public class CustomerController {
      */
     @DeleteMapping("/{id}")
     public ResponseEntity<?> deleteCustomerById(@PathVariable Long id) {
+        Customer customer = customerService.findCustomerById(id);
+        if (customer == null) {
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
         boolean check = customerService.deleteCustomerById(id);
         if (check) {
             return new ResponseEntity<>(HttpStatus.OK);
